+++ examples/07_multistore/src/Person.tsx
@@ -1,0 +1,46 @@
+import * as React from 'react';
+
+import { useReduxDispatch, useReduxState } from 'react-hooks-easy-redux';
+
+import { Action, State } from './state';
+
+const Person = () => {
+  const state = useReduxState<State>();
+  const dispatch = useReduxDispatch<Action>();
+  return (
+    <div>
+      <div>
+        First Name:
+        <input
+          value={state.person.firstName}
+          onChange={(event) => {
+            const firstName = event.target.value;
+            dispatch({ firstName, type: 'setFirstName' });
+          }}
+        />
+      </div>
+      <div>
+        Last Name:
+        <input
+          value={state.person.lastName}
+          onChange={(event) => {
+            const lastName = event.target.value;
+            dispatch({ lastName, type: 'setLastName' });
+          }}
+        />
+      </div>
+      <div>
+        Age:
+        <input
+          value={state.person.age}
+          onChange={(event) => {
+            const age = Number(event.target.value) || 0;
+            dispatch({ age, type: 'setAge' });
+          }}
+        />
+      </div>
+    </div>
+  );
+};
+
+export default Person;